comment 예외 처리

+++ django_app/book/apis/comment.py
@@ -1,8 +1,8 @@
-from rest_framework import generics
+from rest_framework import generics, exceptions
 from rest_framework import status
 from rest_framework.response import Response
 
-from book.models import BookComment
+from book.models import BookComment, MyBook
 from book.serializer import BookCommentSerializer
 
 __all__ = (
@@ -10,27 +10,58 @@ __all__ = (
 )
 
 
-class Comment(generics.GenericAPIView):
+class Comment(generics.GenericAPIView, ):
     queryset = BookComment.objects.all()
     serializer_class = BookCommentSerializer
+    permission_classes = ()
+
+    def get(self, request):
+        if request.auth:
+            user_id = request.GET["user_id"]
+            if user_id:
+                book_list = MyBook.objects.filter(user_id=user_id)
+                comment_list = []
+                for mybook in book_list:
+                    comment, _ = BookComment.objects.get_or_create(mybook_id=mybook.pk)
+                    comment_list.append(comment)
+                serializer = BookCommentSerializer(comment_list, many=True)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                raise exceptions.ParseError({"ios_error_code": 4003, "detail": "Invalid user_id"})
+        else:
+            raise exceptions.NotAuthenticated()
 
     def post(self, request):
-        mybook_id = self.request.data.get('mybook_id')
-        comment = self.request.data.get('comment')
-        defaults = {
-            'comment': comment
-        }
-        q, _ = BookComment.objects.update_or_create(
-            mybook_id=mybook_id,
-            defaults=defaults
-        )
+        if request.auth:
+            mybook_id = self.request.data.get('mybook_id')
+            comment = self.request.data.get('comment')
+            mybook = MyBook.objects.get(id=mybook_id)
+            if mybook.user == request.user:
 
-        serializer = self.get_serializer(q)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+                defaults = {
+                    'comment': comment
+                }
+                q, _ = BookComment.objects.update_or_create(
+                    mybook_id=mybook_id,
+                    defaults=defaults
+                )
+
+                serializer = self.get_serializer(q)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                raise exceptions.ParseError({"detail": "This is not an account on this soobook"})
+        else:
+            raise exceptions.NotAuthenticated()
 
     def delete(self, request):
-        comment_id = self.request.data.get('comment_id')
-        q = super().get_queryset().get(id=comment_id)
-        if q:
-            q.delete()
-            return Response({"detail": "Successfully deleted."}, status=status.HTTP_200_OK)
+        if request.auth:
+            comment_id = self.request.data.get('comment_id')
+            if comment_id:
+                q = super().get_queryset().get(id=comment_id)
+                if q:
+                    q.delete()
+                    return Response({"detail": "Successfully deleted."}, status=status.HTTP_200_OK)
+            else:
+                raise exceptions.ParseError({"detail": "This is a nonexistent comment"})
+        else:
+            raise exceptions.NotAuthenticated()
